New utils.py. Some reformatting of app.py
